+++ common/src/main/java/com/kwwsyk/endinv/common/mixin/ScreenMixin.java
@@ -1,0 +1,13 @@
+package com.kwwsyk.endinv.common.mixin;
+
+import net.minecraft.client.gui.components.events.GuiEventListener;
+import net.minecraft.client.gui.screens.Screen;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(Screen.class)
+public interface ScreenMixin {
+
+    @Invoker("removeWidget")
+    void endinv$invokeRemoveWidget(GuiEventListener widget);
+}